refactor: renames "index.ts" to "exports.ts" in "library/Base64/CharacterSequence"
