+++ src/library/Base64/exports.ts
@@ -1,5 +1,5 @@
 import Base64_Alphabet from './Alphabet';
-import * as Base64_CharacterSequence from './CharacterSequence';
+import * as Base64_CharacterSequence from './CharacterSequence/exports';
 
 const Base64_bitWidth = Math.log2(Base64_Alphabet.length);
